local/requirements.txt
# Lokale Development Dependencies (lightweight!)
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
tqdm>=4.66.0

# Optional: für lokale Tests mit kleinen Modellen
# torch>=2.1.0
# diffusers>=0.25.0
